Django_AuthApi1_DRF_Back/requirements.txt
asgiref==3.7.2
distlib==0.3.7
Django==4.2.7
django-cors-headers==4.2.0
django-dotenv==1.4.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
filelock==3.12.2
mysql-connector==2.2.9
Pillow==10.0.0
platformdirs==3.9.1
PyJWT==2.8.0
pytz==2023.3
sqlparse==0.4.4
tzdata==2023.3

